+++ include/utils.hpp
@@ -10,7 +10,7 @@ namespace PijersiEngine::Utils
     std::vector<std::string> split(std::string str, std::string token = " ");
     std::string strip(std::string str);
 
-    void doubleSort(int16_t* scores, size_t* indices, size_t nMoves);
+    void doubleSort(int64_t* scores, size_t* indices, size_t nMoves);
     void sortPrincipalVariation(std::vector<uint32_t>& moves, uint32_t principalVariation);
 }
